+++ src/main/java/pl/edu/pw/javaee/store/service/CartItemService.java
@@ -11,5 +11,5 @@ public interface CartItemService {
     void addCartItem(CartItem cartItem);
     void removeCartItem(CartItem cartItem);
     void removeAllCartItems(Cart cart);
-    CartItem getCartItemByProductId(int productId);
+    CartItem getCartItemByProductId(long productId);
 }
