+++ src/main/java/pl/edu/pw/javaee/store/service/CustomerService.java
@@ -10,7 +10,7 @@ import java.util.List;
  */
 public interface CustomerService {
     void addCustomer(Customer customer);
-    Customer getCustomerById (int customerId);
+    Customer getCustomerById (long customerId);
     List<Customer> getAllCustomers();
     Customer getCustomerByUsername(String username);
 }
